Move record handling logic out Bolt\Controller\Backend\Records
