+++ src/Application.php
@@ -343,6 +343,7 @@ class Application extends Silex\Application
             ->register(new Silex\Provider\ServiceControllerServiceProvider()) // must be after Routing
             ->register(new Provider\PermissionsServiceProvider())
             ->register(new Provider\StorageServiceProvider())
+            ->register(new Provider\RecordModifierServiceProvider()) // Temporary
             ->register(new Provider\AuthenticationServiceProvider())
             ->register(new Provider\UsersServiceProvider())
             ->register(new Provider\CacheServiceProvider())
