+++ src/Controller/Backend/Records.php
@@ -10,7 +10,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
- * Backend controller for record manipulation rotues.
+ * Backend controller for record manipulation routes.
  *
  * Prior to v2.3 this functionality primarily existed in the monolithic
  * Bolt\Controllers\Backend class.
@@ -41,10 +41,6 @@ class Records extends BackendBase
             ->bind('relatedto')
             ->assert('id', '\d*');
     }
-
-    /*
-     * Routes
-     */
 
     /**
      * Delete a record.
@@ -103,11 +99,31 @@ class Records extends BackendBase
 
         // Save the POSTed record
         if ($request->isMethod('POST')) {
-            return $this->handleSaveRequest($request, $contenttype, $id, $new);
+            $formValues = $request->request->all();
+            $returnTo = $request->get('returnto');
+            $editReferrer = $request->get('editreferrer');
+
+            return $this->app['storage.record_modifier']->handleSaveRequest($formValues, $contenttype, $id, $new, $returnTo, $editReferrer);
+        }
+
+        if ($new) {
+            $content = $this->app['storage']->getEmptyContent($contenttypeslug);
+        } else {
+            $content = $this->getContent($contenttypeslug, array('id' => $id));
+
+            if (empty($content)) {
+                // Record not found, advise and redirect to the dashboard
+                $this->addFlash('error', Trans::__('contenttypes.generic.not-existing', array('%contenttype%' => $contenttypeslug)));
+
+                return $this->redirectToRoute('dashboard');
+            }
         }
 
         // We're doing a GET
-        return $this->handleEditRequest($contenttype, $id, $new);
+        $duplicate = $request->query->get('duplicate', false);
+        $context = $this->app['storage.record_modifier']->handleEditRequest($content, $contenttype, $id, $new, $duplicate);
+
+        return $this->render('editcontent/editcontent.twig', $context);
     }
 
     /**
@@ -290,10 +306,6 @@ class Records extends BackendBase
         return $this->render('relatedto/relatedto.twig', $context);
     }
 
-    /*
-     * actionEditContent() Helper Functions
-     */
-
     /**
      * Check that the user has a valid GSRF token and the required access control
      * to action the record.
@@ -352,423 +364,5 @@ class Records extends BackendBase
         if (strpos($tmpreferrer, '/overview/') !== false || ($tmpreferrer === $this->app['resources']->getUrl('bolt'))) {
             $this->app['twig']->addGlobal('editreferrer', $tmpreferrer);
         }
-    }
-
-    /**
-     * Do the save for a POSTed record.
-     *
-     * @param Request $request
-     * @param array   $contenttype The contenttype data
-     * @param integer $id          The record ID
-     * @param boolean $new         If TRUE this is a new record
-     *
-     * @return Response
-     */
-    private function handleSaveRequest(Request $request, array $contenttype, $id, $new)
-    {
-        $contenttypeslug = $contenttype['slug'];
-
-        // If we have an ID now, this is an existing record
-        if ($id) {
-            $content = $this->getContent($contenttypeslug, array('id' => $id, 'status' => '!'));
-            $oldStatus = $content['status'];
-        } else {
-            $content = $this->app['storage']->getContentObject($contenttypeslug);
-            $oldStatus = '';
-        }
-
-        // Don't allow spoofing the $id.
-        if (!empty($content['id']) && $id != $content['id']) {
-            $this->addFlash('error', "Don't try to spoof the id!");
-
-            return $this->redirectToRoute('dashboard');
-        }
-
-        // Ensure all fields have valid values
-        $requestAll = $this->setSuccessfulControlValues($request, $contenttype['fields']);
-
-        // To check whether the status is allowed, we act as if a status
-        // *transition* were requested.
-        $content->setFromPost($requestAll, $contenttype);
-        $newStatus = $content['status'];
-
-        $statusOK = $this->app['users']->isContentStatusTransitionAllowed($oldStatus, $newStatus, $contenttypeslug, $id);
-        if ($statusOK) {
-            // Save the record
-            return $this->saveContentRecord($request, $contenttype, $content, $new);
-        } else {
-            $this->addFlash('error', Trans::__('contenttypes.generic.error-saving', array('%contenttype%' => $contenttypeslug)));
-            $this->app['logger.system']->error('Save error: ' . $content->getTitle(), array('event' => 'content'));
-        }
-    }
-
-    /**
-     * Commit the record to the database.
-     *
-     * @param Request $request
-     * @param array   $contenttype
-     * @param Content $content
-     * @param boolean $new
-     *
-     * @return Response
-     */
-    private function saveContentRecord(Request $request, array $contenttype, Content $content, $new)
-    {
-        // Get the associated record change comment
-        $comment = $request->request->get('changelog-comment');
-
-        // Save the record
-        $id = $this->app['storage']->saveContent($content, $comment);
-
-        // Log the change
-        if ($new) {
-            $this->addFlash('success', Trans::__('contenttypes.generic.saved-new', array('%contenttype%' => $contenttype['slug'])));
-            $this->app['logger.system']->info('Created: ' . $content->getTitle(), array('event' => 'content'));
-        } else {
-            $this->addFlash('success', Trans::__('contenttypes.generic.saved-changes', array('%contenttype%' => $contenttype['slug'])));
-            $this->app['logger.system']->info('Saved: ' . $content->getTitle(), array('event' => 'content'));
-        }
-
-        /*
-         * We now only get a returnto parameter if we are saving a new
-         * record and staying on the same page, i.e. "Save {contenttype}"
-         */
-        if ($this->app['request']->get('returnto')) {
-            $returnto = $this->app['request']->get('returnto');
-
-            if ($returnto === 'new') {
-                return $this->redirectToRoute('editcontent', array('contenttypeslug' => $contenttype['slug'], 'id' => $id));
-            } elseif ($returnto == 'saveandnew') {
-                return $this->redirectToRoute('editcontent', array('contenttypeslug' => $contenttype['slug']));
-            } elseif ($returnto === 'ajax') {
-                return $this->createJsonUpdate($contenttype, $id);
-            }
-        }
-
-        // No returnto, so we go back to the 'overview' for this contenttype.
-        // check if a pager was set in the referrer - if yes go back there
-        $editreferrer = $this->app['request']->get('editreferrer');
-        if ($editreferrer) {
-            return $this->redirect($editreferrer);
-        } else {
-            return $this->redirectToRoute('overview', array('contenttypeslug' => $contenttype['slug']));
-        }
-    }
-
-    /**
-     * Add successful control values to request values, and do needed corrections.
-     *
-     * @see http://www.w3.org/TR/html401/interact/forms.html#h-17.13.2
-     *
-     * @param Request $request
-     * @param array   $fields
-     *
-     * @return array
-     */
-    private function setSuccessfulControlValues(Request $request, $fields)
-    {
-        $formValues = $request->request->all();
-
-        foreach ($fields as $key => $values) {
-            if (isset($formValues[$key])) {
-                switch ($values['type']) {
-                    case 'float':
-                        // We allow ',' and '.' as decimal point and need '.' internally
-                        $formValues[$key] = str_replace(',', '.', $formValues[$key]);
-                        break;
-                }
-            } else {
-                switch ($values['type']) {
-                    case 'select':
-                        if (isset($values['multiple']) && $values['multiple'] === true) {
-                            $formValues[$key] = array();
-                        }
-                        break;
-
-                    case 'checkbox':
-                        $formValues[$key] = 0;
-                        break;
-                }
-            }
-        }
-
-        return $formValues;
-    }
-
-    /**
-     * Build a valid AJAX response for in-place saves that account for pre/post
-     * save events.
-     *
-     * @param array   $contenttype
-     * @param integer $id
-     *
-     * @return JsonResponse
-     */
-    private function createJsonUpdate($contenttype, $id)
-    {
-        /*
-         * Flush any buffers from saveConent() dispatcher hooks
-         * and make sure our JSON output is clean.
-         *
-         * Currently occurs due to exceptions being generated in the dispatchers
-         * in \Bolt\Storage::saveContent()
-         *     StorageEvents::PRE_SAVE
-         *     StorageEvents::POST_SAVE
-         */
-        Response::closeOutputBuffers(0, false);
-
-        // Get our record after POST_SAVE hooks are dealt with and return the JSON
-        $content = $this->getContent($contenttype['slug'], array('id' => $id, 'returnsingle' => true, 'status' => '!'));
-
-        $val = array();
-
-        foreach ($content->values as $key => $value) {
-            // Some values are returned as \Twig_Markup and JSON can't deal with that
-            if (is_array($value)) {
-                foreach ($value as $subkey => $subvalue) {
-                    if (gettype($subvalue) == 'object' && get_class($subvalue) == 'Twig_Markup') {
-                        $val[$key][$subkey] = (string) $subvalue;
-                    }
-                }
-            } else {
-                $val[$key] = $value;
-            }
-        }
-
-        if (isset($val['datechanged'])) {
-            $val['datechanged'] = date_format(new \DateTime($val['datechanged']), 'c');
-        }
-
-        $lc = localeconv();
-        foreach ($contenttype['fields'] as $key => $values) {
-            switch ($values['type']) {
-                case 'float':
-                    // Adjust decimal point dependent on locale
-                    if ($lc['decimal_point'] === ',') {
-                        $val[$key] = str_replace('.', ',', $val[$key]);
-                    }
-                    break;
-            }
-        }
-
-        // Unset flashbag for ajax
-        $this->app['logger.flash']->clear();
-
-        return $this->json($val);
-    }
-
-    /**
-     * Do the edit rendering for a record.
-     *
-     * @param array   $contenttype The contenttype data
-     * @param integer $id          The record ID
-     * @param boolean $new         If TRUE this is a new record
-     *
-     * @return \Bolt\Response\BoltResponse
-     */
-    private function handleEditRequest(array $contenttype, $id, $new)
-    {
-        $contenttypeslug = $contenttype['slug'];
-
-        if ($new) {
-            $content = $this->app['storage']->getEmptyContent($contenttypeslug);
-        } else {
-            $content = $this->getContent($contenttypeslug, array('id' => $id));
-
-            if (empty($content)) {
-                // Record not found, advise and redirect to the dashboard
-                $this->addFlash('error', Trans::__('contenttypes.generic.not-existing', array('%contenttype%' => $contenttypeslug)));
-
-                return $this->redirectToRoute('dashboard');
-            }
-        }
-
-        $oldStatus = $content['status'];
-        $allStatuses = array('published', 'held', 'draft', 'timed');
-        $allowedStatuses = array();
-        foreach ($allStatuses as $status) {
-            if ($this->app['users']->isContentStatusTransitionAllowed($oldStatus, $status, $contenttypeslug, $id)) {
-                $allowedStatuses[] = $status;
-            }
-        }
-
-        // For duplicating a record, clear base field values
-        if ($duplicate = $this->app['request']->query->get('duplicate')) {
-            $content->setValues(array(
-                'id'            => '',
-                'slug'          => '',
-                'datecreated'   => '',
-                'datepublish'   => '',
-                'datedepublish' => null,
-                'datechanged'   => '',
-                'username'      => '',
-                'ownerid'       => '',
-            ));
-
-            $this->addFlash('info', Trans::__('contenttypes.generic.duplicated-finalize', array('%contenttype%' => $contenttypeslug)));
-        }
-
-        // Set the users and the current owner of this content.
-        if ($new || $duplicate) {
-            // For brand-new and duplicated items, the creator becomes the owner.
-            $contentowner = $this->getUser();
-        } else {
-            // For existing items, we'll just keep the current owner.
-            $contentowner = $this->getUser($content['ownerid']);
-        }
-
-        // Test write access for uploadable fields
-        $contenttype['fields'] = $this->setCanUpload($contenttype['fields']);
-        if ((!empty($content['templatefields'])) && (!empty($content['templatefields']->contenttype['fields']))) {
-            $content['templatefields']->contenttype['fields'] = $this->setCanUpload($content['templatefields']->contenttype['fields']);
-        }
-
-        // Determine which templates will result in templatefields
-        $templateFieldTemplates = $this->getTempateFieldTemplates($contenttype, $content);
-
-        // Information flags about what the record contains
-        $info = array(
-            'hasIncomingRelations' => is_array($content->relation),
-            'hasRelations'         => isset($contenttype['relations']),
-            'hasTabs'              => $contenttype['groups'] !== false,
-            'hasTaxonomy'          => isset($contenttype['taxonomy']),
-            'hasTemplateFields'    => $content->hasTemplateFields()
-        );
-
-        // Generate tab groups
-        $groups = $this->createGroupTabs($contenttype, $content, $info);
-
-        // Build context for Twig
-        $context = array(
-            'contenttype'    => $contenttype,
-            'content'        => $content,
-            'allowed_status' => $allowedStatuses,
-            'contentowner'   => $contentowner,
-            'fields'         => $this->app['config']->fields->fields(),
-            'fieldtemplates' => $templateFieldTemplates,
-            'can_upload'     => $this->isAllowed('files:uploads'),
-            'groups'         => $groups,
-            'has'            => array(
-                'incoming_relations' => $info['hasIncomingRelations'],
-                'relations'          => $info['hasRelations'],
-                'tabs'               => $info['hasTabs'],
-                'taxonomy'           => $info['hasTaxonomy'],
-                'templatefields'     => $info['hasTemplateFields'],
-            ),
-        );
-
-        // Render
-        return $this->render('editcontent/editcontent.twig', $context);
-    }
-
-    /**
-     * Test write access for uploadable fields
-     *
-     * @param array $fields
-     *
-     * @return array
-     */
-    private function setCanUpload(array $fields)
-    {
-        $filesystem = $this->getFilesystemManager()->getFilesystem();
-
-        foreach ($fields as &$values) {
-            if (isset($values['upload'])) {
-                $values['canUpload'] = $filesystem->has($values['upload']) && $filesystem->getVisibility($values['upload']);
-            } else {
-                $values['canUpload'] = true;
-            }
-        }
-
-        return $fields;
-    }
-
-    /**
-     * Determine which templates will result in templatefields
-     *
-     * @param array   $contenttype
-     * @param Content $content
-     *
-     * @return array
-     */
-    private function getTempateFieldTemplates(array $contenttype, Content $content)
-    {
-        $templateFieldTemplates = array();
-
-        if ($templateFieldsConfig = $this->app['config']->get('theme/templatefields')) {
-            $templateFieldTemplates = array_keys($templateFieldsConfig);
-            // Special case for default template
-            $toRepair = array();
-            foreach ($contenttype['fields'] as $name => $field) {
-                if ($field['type'] == 'templateselect' && !empty($content->values[$name])) {
-                    $toRepair[$name] = $content->values[$name];
-                    $content->setValue($name, '');
-                }
-            }
-            if ($content->hasTemplateFields()) {
-                $templateFieldTemplates[] = '';
-            }
-
-            foreach ($toRepair as $name => $value) {
-                $content->setValue($name, $value);
-            }
-        }
-
-        return $templateFieldTemplates;
-    }
-
-    /**
-     * Generate tab groups
-     *
-     * @param array   $contenttype
-     * @param Content $content
-     * @param array   $info
-     *
-     * @return array
-     */
-    private function createGroupTabs(array $contenttype, Content $content, $info)
-    {
-        $groups = array();
-        $groupIds = array();
-
-        $addGroup = function ($group, $label) use (&$groups, &$groupIds) {
-            $nr = count($groups) + 1;
-            $id = rtrim('tab-' . Slugify::create()->slugify($group), '-');
-            if (isset($groupIds[$id]) || $id === 'tab') {
-                $id .= '-' . $nr;
-            }
-            $groups[$group] = array(
-                'label'     => $label,
-                'id'        => $id,
-                'is_active' => $nr === 1,
-            );
-            $groupIds[$id] = 1;
-        };
-
-        foreach ($contenttype['groups'] ? $contenttype['groups'] : array('ungrouped') as $group) {
-            if ($group === 'ungrouped') {
-                $addGroup($group, Trans::__('contenttypes.generic.group.ungrouped'));
-            } elseif ($group !== 'meta' && $group !== 'relations' && $group !== 'taxonomy') {
-                $default = array('DEFAULT' => ucfirst($group));
-                $key = array('contenttypes', $contenttype['slug'], 'group', $group);
-                $addGroup($group, Trans::__($key, $default));
-            }
-        }
-
-        if ($info['hasRelations'] || $info['hasIncomingRelations']) {
-            $addGroup('relations', Trans::__('contenttypes.generic.group.relations'));
-        }
-
-        if ($info['hasTaxonomy'] || (is_array($contenttype['groups']) && in_array('taxonomy', $contenttype['groups']))) {
-            $addGroup('taxonomy', Trans::__('contenttypes.generic.group.taxonomy'));
-        }
-
-        if ($info['hasTemplateFields'] || (is_array($contenttype['groups']) && in_array('template', $contenttype['groups']))) {
-            $addGroup('template', Trans::__('Template'));
-        }
-
-        $addGroup('meta', Trans::__('contenttypes.generic.group.meta'));
-
-        return $groups;
     }
 }
